requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.5.2               # via kombu
appnope==0.1.0            # via ipython
autopep8==1.5.2           # via -r requirements.in
backcall==0.1.0           # via ipython
beautifulsoup4==4.9.0     # via bs4
billiard==3.6.3.0         # via celery
brotlipy==0.7.0           # via -r requirements.in
bs4==0.0.1                # via -r requirements.in
celery[redis]==4.4.2      # via -r requirements.in
certifi==2020.4.5.1       # via requests, sentry-sdk
cffi==1.14.0              # via brotlipy
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via odfpy
diff-match-patch==20181111  # via django-import-export
dj-database-url==0.5.0    # via -r requirements.in
django-import-export==2.1.0  # via -r requirements.in
django-js-reverse==0.9.1  # via -r requirements.in
django-log-request-id==1.5.0  # via -r requirements.in
django-model-utils==4.0.0  # via -r requirements.in
django-webpack-loader==0.7.0  # via -r requirements.in
django==2.2.13            # via -r requirements.in, django-import-export, django-js-reverse, django-log-request-id, django-model-utils, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
dnspython==1.16.0         # via -r requirements.in
et-xmlfile==1.0.1         # via openpyxl
feedparser==5.2.1         # via -r requirements.in
flake8==3.8.1             # via -r requirements.in
fuzzywuzzy[speedup]==0.18.0  # via -r requirements.in
geographiclib==1.50       # via geopy
geopy==1.22.0             # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.9                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==7.14.0           # via -r requirements.in
jdcal==1.4.1              # via openpyxl
jedi==0.17.0              # via ipython
kiwisolver==1.2.0         # via matplotlib
kombu==4.6.8              # via celery
markuppy==1.14            # via tablib
matplotlib==3.2.1         # via -r requirements.in
mccabe==0.6.1             # via flake8
numpy==1.18.4             # via matplotlib, pandas
odfpy==1.4.1              # via tablib
openpyxl==3.0.3           # via tablib
pandas==1.0.3             # via -r requirements.in
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==3.0.5     # via ipython
psutil==5.7.0             # via -r requirements.in
psycopg2==2.8.5           # via -r requirements.in
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.6.0        # via autopep8, flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pygments==2.6.1           # via ipython
pymongo==3.10.1           # via -r requirements.in
pyparsing==2.4.7          # via matplotlib
python-dateutil==2.8.1    # via matplotlib, pandas
python-decouple==3.3      # via -r requirements.in
python-http-client==3.2.7  # via sendgrid
python-levenshtein==0.12.0  # via fuzzywuzzy
pytz==2020.1              # via celery, django, pandas
pyyaml==5.3.1             # via tablib
redis==3.5.1              # via celery
requests==2.23.0          # via -r requirements.in, sodapy
sendgrid==6.3.1           # via -r requirements.in
sentry-sdk==0.14.4        # via -r requirements.in
six==1.14.0               # via cycler, python-dateutil, traitlets
sodapy==2.1.0             # via -r requirements.in
soupsieve==2.0            # via beautifulsoup4
sqlparse==0.3.1           # via django
tablib[html,ods,xls,xlsx,yaml]==1.1.0  # via django-import-export
traitlets==4.3.3          # via ipython
urllib3==1.25.9           # via requests, sentry-sdk
vine==1.3.0               # via amqp, celery
wcwidth==0.1.9            # via prompt-toolkit
wheel==0.34.2             # via -r requirements.in
whitenoise==5.0.1         # via -r requirements.in
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
yattag==1.13.2            # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
